renamed icon to icon wrapper
