+++ blkn/src/Button/__tests__/icon.test.js
@@ -1,10 +1,10 @@
 import * as React from "react";
 import { shallow } from "enzyme";
 import { Money } from "@kiwicom/icons";
-import Icon from "../Icon";
+import IconWrapper from "../IconWrapper";
 
 describe("Icon", () => {
-  const component = shallow(<Icon icon={Money} />);
+  const component = shallow(<IconWrapper icon={Money} />);
 
   it("should be moiunted with svg element", () => {
     expect(component.find("svg").exists()).toBe(true);
