+++ blkn/src/index.js
@@ -1,7 +1,7 @@
 // @flow
 
 export { default as Button } from "./Button";
-export { default as Icon } from "./Button/Icon";
+export { default as IconWrapper } from "./Button/IconWrapper";
 export { default as InputText } from "./InputText";
 export { default as Checkbox } from "./Checkbox";
 export { default as Typography } from "./Typography";
